mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sheetstorm-mcp"
version = "0.1.0"
description = "MCP (Model Context Protocol) server for SheetStorm incident response platform"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "SheetStorm Team" },
]
keywords = ["mcp", "incident-response", "sheetstorm", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.scripts]
sheetstorm-mcp = "sheetstorm_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["sheetstorm_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
